+++ packages/cli/core/src/lib/utils/process.spec.ts
@@ -1,25 +1,19 @@
 import {EventEmitter} from 'events';
-import {ChildProcess} from 'child_process';
-import {spawnProcess} from './process';
+import type {ChildProcess} from 'child_process';
+import {handleForkedProcess, spawnProcess} from './process';
 import {fancyIt} from '@coveo/cli-commons-dev/testUtils/it';
+jest.mock('child_process');
+import {spawn} from 'child_process';
 
-jest.mock('child_process', () => ({
-  spawn: jest
-    .fn()
-    .mockImplementationOnce(() => {
+describe('spawnProcess', () => {
+  const mockedSpawn = jest.mocked(spawn);
+  fancyIt()('should resolve with a success exit code', () => {
+    mockedSpawn.mockImplementationOnce(() => {
       const emitter = new EventEmitter();
       process.nextTick(() => emitter.emit('close', 0));
       return emitter as ChildProcess;
-    })
-    .mockImplementationOnce(() => {
-      const emitter = new EventEmitter();
-      process.nextTick(() => emitter.emit('close', 1));
-      return emitter as ChildProcess;
-    }),
-}));
+    });
 
-describe('spawnProcess', () => {
-  fancyIt()('should resolves with a success exit code', () => {
     const command = 'some valid command';
     const args = ['-valid', 'option'];
 
@@ -27,9 +21,37 @@ describe('spawnProcess', () => {
   });
 
   fancyIt()('should reject', () => {
+    mockedSpawn.mockImplementationOnce(() => {
+      const emitter = new EventEmitter();
+      process.nextTick(() => emitter.emit('close', 1));
+      return emitter as ChildProcess;
+    });
+
     const command = 'invalid commande';
     const args = ['-foo', 'bar'];
 
     return expect(spawnProcess(command, args)).rejects.toEqual(1);
+  });
+
+  describe('handleForkedProcess', () => {
+    it('should resolve the returned promise when the subprocess exits', async () => {
+      const fakeSubProcess = (() => {
+        const emitter = new EventEmitter();
+        process.nextTick(() => emitter.emit('exit', 0));
+        return emitter as ChildProcess;
+      })();
+
+      await expect(handleForkedProcess(fakeSubProcess)).resolves.not.toThrow();
+    });
+
+    it('should reject the returned promise when the subprocess sends a message', async () => {
+      const fakeSubProcess = (() => {
+        const emitter = new EventEmitter();
+        process.nextTick(() => emitter.emit('message', 'potato'));
+        return emitter as ChildProcess;
+      })();
+
+      await expect(handleForkedProcess(fakeSubProcess)).rejects.toBe('potato');
+    });
   });
 });